게시물삭제,게시물수정,findByTitle 로직수정 및 추가

+++ src/main/java/com/mysite/sbb/Controller/ArticleController.java
@@ -7,13 +7,21 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Controller
-@RequestMapping("usr/article")
+@RequestMapping("/usr/article")
 public class ArticleController {
     @Autowired
     private ArticleRepository articleRepository;
+
+    @RequestMapping("/test")
+    @ResponseBody
+    public String testFunc() {
+        return "test";
+    }
 
     @RequestMapping("/list")
     @ResponseBody
@@ -24,20 +32,33 @@ public class ArticleController {
     @RequestMapping("/detail")
     @ResponseBody
     public Article showDetail(Long id) {
-        Article article = articleRepository.findById(id).get();
-        return article;
+        Optional<Article> article = articleRepository.findById(id);
+        return article.orElse(null);
     }
 
-    @RequestMapping("/delete")
+    @RequestMapping("/doDelete")
     @ResponseBody
-    public String DoArticleDelete(Long id) {
+    public String doDelete(Long id) {
+        if(articleRepository.existsById(id) == false) {
+            return "%d번 게시물은 이미 삭제되었거나 존재하지 않습니다.".formatted(id);
+        }
+
         articleRepository.deleteById(id);
 
         return "%d번 게시물이 삭제되었습니다.".formatted(id);
+
     }
-    @RequestMapping("/Modify")
+
+    @RequestMapping("/findByTitle")
     @ResponseBody
-    public Article DoArticleModify(Long id, String title, String body) {
+    public List<Article> findByTitle(String title) {
+        List<Article> articles = articleRepository.findByTitle(title);
+        return articles;
+    }
+
+    @RequestMapping("/doModify")
+    @ResponseBody
+    public Article doModify(Long id, String title, String body) {
         Article article = articleRepository.findById(id).get();
 
         if(title != null) {
@@ -46,6 +67,9 @@ public class ArticleController {
         if(body != null) {
             article.setBody(body);
         }
+
+        article.setUpdateDate(LocalDateTime.now());
+
         articleRepository.save(article);
         return article;
     }
